Extract AuthUser identity into lib/shared
The authenticated user identity is the one business type used across
multiple features (auth, profile). Move it out of the auth feature into a
new lib/shared/ module that holds cross-feature domain vocabulary, distinct
from lib/core/ (infrastructure).

- Move AuthUser to lib/shared/domain/entities/auth_user.dart
- Repoint all 9 lib importers and 3 test files to the shared path
- Document the core vs shared vs features split (and the rule-of-three gate
  for promoting types into shared) in CLAUDE.md

Dependency direction is now features -> shared -> core. No behavior change;
flutter analyze clean, all 289 tests pass.
