+++ lib/features/auth/domain/usecases/restore_session.dart
@@ -2,7 +2,7 @@ import 'package:injectable/injectable.dart';
 
 import '../../../../core/usecases/use_case.dart';
 import '../../../../core/utils/result.dart';
-import '../entities/auth_user.dart';
+import '../../../../shared/domain/entities/auth_user.dart';
 import '../repositories/auth_repository.dart';
 
 @injectable
