+++ lib/shared/domain/entities/auth_user.dart
